+++ test/index.js
@@ -31,6 +31,12 @@ test('create a service when the url has a trailing slash', function (t) {
   t.end()
 })
 
+test('override layer with passed in option', function (t) {
+  var serv = new FeatureService('http://koop.whateva.com/FeatureServer/1', {layer: '3'})
+  t.equal(serv.layer.toString(), '3')
+  t.end()
+})
+
 test('get the objectId', function (t) {
   var oid = service.getObjectIdField(layerInfo)
   t.equal(oid, 'ESRI_OID')